chore(deps): upgrade rusqlite 0.31→0.38, thiserror 1→2, rhai min 1.17→1.24
rusqlite 0.38 dropped ToSql for usize; fix purge_if_needed to cast
keep_last as i64 before passing to execute().

thiserror 2.x was already pulled in transitively; aligning the
workspace spec eliminates the duplicate copy in the lock file.

rhai 1.24 was already locked; updating the minimum spec to match.

All 31 tests pass.

+++ krillnotes-core/src/core/operation_log.rs
@@ -67,7 +67,7 @@ impl OperationLog {
                     "DELETE FROM operations WHERE id NOT IN (
                         SELECT id FROM operations ORDER BY id DESC LIMIT ?
                     )",
-                    [keep_last],
+                    [keep_last as i64],
                 )?;
             }
             PurgeStrategy::WithSync { retention_days } => {
